requirements.txt
fastapi==0.115.12
SQLAlchemy==2.0.40
uvicorn==0.34.0
pandas>=2.2.0
numpy>=2.2.0
pytz
celery>=5.5.1
psycopg2>=2.9.10
python-dotenv
pydantic